requirements.txt
Django==3.2
djangorestframework==3.12.4
django-cors-headers==3.7.0
django-health-check==3.16.4
django-oauth-toolkit==1.5.0
django-storages[google]==1.11.1
django-filter==2.4.0

Pillow==8.2.0
stripe==2.56.0
psycopg2==2.8.6
python-dotenv==0.17.0
google-cloud-core==1.6.0
google-cloud-storage==1.38.0